Move default aggregator YAML regression test to docs/
Addresses review feedback to keep the docs regression test alongside the
generated YAML it validates rather than in fme/ace/.
